Fix errors. Add student update function
