+++ src/services/user.service.js
@@ -66,6 +66,19 @@ const updateUserById = async (userId, updateBody) => {
   return user;
 };
 
+const addStudentInformation = async (userId, updateBody) => {
+  const user = await getUserById(userId);
+  if (!user) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'User not found');
+  }
+  if (!updateBody.age && !updateBody.hobby) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'Email already taken');
+  }
+  Object.assign(user, updateBody);
+  await user.save();
+  return user;
+};
+
 /**
  * Delete user by id
  * @param {ObjectId} userId
@@ -86,5 +99,6 @@ module.exports = {
   getUserById,
   getUserByEmail,
   updateUserById,
-  deleteUserById
+  deleteUserById,
+  addStudentInformation,
 };
